Allow for different transaction step checks
